Refactor common classes in model to model.commons
