+++ src/main/java/seedu/address/model/contact/Contact.java
@@ -7,6 +7,8 @@ import java.util.HashSet;
 import java.util.Objects;
 import java.util.Set;
 
+import seedu.address.model.commons.Address;
+import seedu.address.model.commons.Name;
 import seedu.address.model.tag.Tag;
 
 /**
