+++ src/main/java/seedu/address/model/task/Location.java
@@ -2,7 +2,7 @@ package seedu.address.model.task;
 
 import static java.util.Objects.requireNonNull;
 
-import seedu.address.model.contact.Name;
+import seedu.address.model.commons.Name;
 
 public class Location {
     public final String location;
